Auto-detect internal packages

+++ scripts/build.sh
@@ -4,20 +4,22 @@ set -e
 
 ./scripts/install-build-deps.sh
 
-PACKAGES=( "crypto" "schnorrkel" )
-
 echo "*** Building packages"
 cd packages
 
+PACKAGES=( $(ls -1d *) )
+
 for PKG in "${PACKAGES[@]}"; do
-  echo "*** Building wasm-$PKG"
-  cd wasm-$PKG
+  if [ -f "$PKG/package.json" ]; then
+    echo "*** Building $PKG"
+    cd $PKG
 
-  ./build.sh
-  rm -rf build/*-e build/package.json build/README.md
-  ls -al build
+    ./build.sh
+    rm -rf build/*-e build/package.json build/README.md
+    ls -al build
 
-  cd ..
+    cd ..
+  fi
 done
 
 cd ..
